refactor: TrtLLM WideEP MoE family owns perf data (AIC-536 / Phase 4.2)
Phase 4 — moves ``TrtLLMWideEPMoE`` and ``TrtLLMWideEPMoEDispatch`` out
of ``operations/_legacy.py`` into the existing ``operations/moe.py``
(co-located with MoE + MoEDispatch from ISSUE-12).

``TrtLLMWideEPMoE`` owns ``_wideep_moe_compute_data`` (TRT-LLM-only,
``None`` on other backends). Pulls the kernel-selection helper onto
the class:

- ``_select_kernel`` (formerly ``PerfDatabase._select_moe_kernel``) —
  picks ``deepgemm`` on Blackwell+fp8_block, otherwise
  ``moe_torch_flow``; falls back to whatever's loaded.

``TrtLLMWideEPMoEDispatch`` owns ``_trtllm_alltoall_data``
(TRT-LLM-only, ``None`` on other backends). Pulls two helpers onto the
class:

- ``_normalize_quant_mode_for_table`` (formerly
  ``PerfDatabase._normalize_alltoall_moe_quant_mode_for_table``) —
  ``fp8_block`` reuses ``fp8`` alltoall tables.
- ``_select_alltoall_kernel`` (formerly
  ``PerfDatabase._select_alltoall_kernel``) — picks NVLinkOneSided /
  NVLinkTwoSided / DeepEP / DeepEPLowLatency / NotEnabled based on
  SM version, ``moe_backend``, inter-node-ness, and topk.

``PerfDatabase.query_wideep_moe_compute`` and ``query_trtllm_alltoall``
reduce to one-line delegations to the new classmethods. The three
helper methods on ``PerfDatabase`` are deleted outright (no other
callers).

The ``if backend == "trtllm":`` legacy init block disappears; both
classes are loaded via the eager
``TrtLLMWideEPMoE.load_data(self)`` /
``TrtLLMWideEPMoEDispatch.load_data(self)`` calls added in
``PerfDatabase.__init__``. ISSUE-16 cleanup retires those eager calls.

Logger note: ``operations/moe.py`` switches to a late-binding
``_LoggerProxy`` over ``_legacy.logger`` so the existing test pattern
``mock.patch("aiconfigurator.sdk.operations._legacy.logger")`` keeps
capturing log calls from the migrated classes. Three
``test_query_debug_logging`` tests were broken by the import-time
``from _legacy import logger`` style; the proxy fixes them without
touching the test surface.
Verification:
- 670 unit tests pass + 1 skipped (same as baseline)
- 38 regression snapshots bit-identical + 6 skipped (same as baseline)
- ruff check + format clean
