+++ src/aiconfigurator/sdk/perf_database.py
@@ -2834,10 +2834,10 @@ class PerfDatabase:
         # (ContextDSAModule.load_data / GenerationDSAModule.load_data handle
         # the load + deepcopy of the raw rows + grid extrapolation).
 
-        # TensorRT-LLM wideep path
-        if backend == "trtllm":
-            self._wideep_moe_compute_data = _load_op_data(PerfDataFilename.wideep_moe_compute)
-            self._trtllm_alltoall_data = _load_op_data(PerfDataFilename.trtllm_alltoall)
+        # TensorRT-LLM wideep path: ``_wideep_moe_compute_data`` and
+        # ``_trtllm_alltoall_data`` are loaded by ``TrtLLMWideEPMoe.load_data`` /
+        # ``TrtLLMWideEPMoeDispatch.load_data`` below (ISSUE-13). Both gate on
+        # ``database.backend == "trtllm"``.
 
         # GEMM / ContextAttention / GenerationAttention own their CSV tables +
         # SOL correction + grid extrapolation. The eager invocations here are
@@ -2865,7 +2865,12 @@ class PerfDatabase:
             WideEPContextMLA,
             WideEPGenerationMLA,
         )
-        from aiconfigurator.sdk.operations.moe import MoE, MoEDispatch
+        from aiconfigurator.sdk.operations.moe import (
+            MoE,
+            MoEDispatch,
+            TrtLLMWideEPMoE,
+            TrtLLMWideEPMoEDispatch,
+        )
 
         GEMM.load_data(self)
         ContextAttention.load_data(self)
@@ -2887,6 +2892,8 @@ class PerfDatabase:
         GenerationDeepSeekV4AttentionModule.load_data(self)
         MoE.load_data(self)
         MoEDispatch.load_data(self)
+        TrtLLMWideEPMoE.load_data(self)
+        TrtLLMWideEPMoEDispatch.load_data(self)
 
         # pre-correction (non-migrated ops; migrated ops apply their own
         # SOL correction during ``load_data``)
@@ -3161,130 +3168,6 @@ class PerfDatabase:
         Check if the number of GPUs is an inter node
         """
         return num_gpus > self.system_spec["node"]["num_gpus_per_node"]
-
-    def _select_alltoall_kernel(
-        self,
-        quant_mode: common.MoEQuantMode,
-        moe_ep_size: int,
-        topk: int,
-        moe_backend: Optional[str] = None,
-    ) -> str:
-        """
-        Automatically select All2All communication method based on GPU architecture,
-        MoE backend type, and configuration.
-
-        Aligned with TensorRT-LLM's per-backend select_alltoall_method_type:
-
-        CutlassFusedMoE / TRTLLMGenFusedMoE (fused_moe_cutlass.py / fused_moe_trtllm_gen.py):
-          - Requires supports_mnnvl() (approximated as SM >= 100)
-          - Returns NVLinkOneSided
-          - Does NOT support DeepEP / DeepEPLowLatency
-
-        WideEPMoE (fused_moe_wide_ep.py):
-          - If supports_mnnvl() -> NVLinkTwoSided
-          - Else if DeepEP feasible -> DeepEP (inter-node) or DeepEPLowLatency (intra-node)
-          - Does NOT support NVLinkOneSided
-
-        DeepGemmFusedMoE / CuteDslFusedMoE:
-          - Always NotEnabled
-
-        Args:
-            quant_mode: MoE quantization mode
-            moe_ep_size: MoE expert parallelism size
-            topk: Number of experts activated per token
-            moe_backend: MoE backend identifier. "wideep" for WideEP path,
-                        "CUTLASS"/"TRTLLM"/None for CutlassFusedMoE/TRTLLMGen,
-                        "DEEPGEMM"/"CUTE_DSL" for backends without AlltoAll.
-
-        Returns:
-            str: The selected kernel_source name, or "NotEnabled" if AlltoAll is not used.
-        """
-        if moe_backend is not None and moe_backend.upper() in {"DEEPGEMM", "CUTE_DSL"}:
-            return "NotEnabled"
-
-        sm_version = self.system_spec["gpu"]["sm_version"]
-        num_gpus_per_node = self.system_spec["node"]["num_gpus_per_node"]
-        is_inter_node = moe_ep_size > num_gpus_per_node
-        is_wideep = moe_backend is not None and moe_backend.upper() == "WIDEEP"
-
-        supports_mnnvl = sm_version >= 100
-
-        if is_wideep:
-            if supports_mnnvl:
-                preferred = "NVLinkTwoSided"
-            else:
-                deepep_feasible = moe_ep_size > 1 and topk <= 8
-                if deepep_feasible and is_inter_node:
-                    preferred = "DeepEP"
-                elif deepep_feasible:
-                    preferred = "DeepEPLowLatency"
-                else:
-                    preferred = "NotEnabled"
-        else:
-            if supports_mnnvl:
-                preferred = "NVLinkOneSided"
-            else:
-                preferred = "NotEnabled"
-
-        if preferred == "NotEnabled":
-            return preferred
-
-        if self._trtllm_alltoall_data:
-            available_kernels = list(self._trtllm_alltoall_data.keys())
-            if preferred in available_kernels:
-                return preferred
-            else:
-                logger.warning(
-                    f"Preferred All2All kernel '{preferred}' not in available kernels {available_kernels}. "
-                    f"Returning preferred anyway; downstream will fall back to HYBRID estimation."
-                )
-
-        return preferred
-
-    def _select_moe_kernel(
-        self,
-        quant_mode: common.MoEQuantMode,
-    ) -> str:
-        """
-        Automatically select MoE computation kernel based on GPU architecture and quantization mode.
-
-        Selection logic (based on TensorRT-LLM's MoEOpSelector.select_op):
-        1. SM >= 100 (Blackwell) with fp8_block -> deepgemm (DeepGemm kernel)
-        2. Otherwise -> moe_torch_flow (Cutlass kernel)
-
-        Args:
-            quant_mode: MoE quantization mode
-
-        Returns:
-            str: The selected kernel_source name
-        """
-        sm_version = self.system_spec["gpu"]["sm_version"]
-        is_blackwell = sm_version >= 100
-
-        # Convert quant_mode to string for comparison if needed
-        quant_mode_str = quant_mode.name if hasattr(quant_mode, "name") else str(quant_mode)
-        is_fp8_block = "fp8_block" in quant_mode_str
-
-        # Preferred kernel based on hardware and quant mode
-        if is_blackwell and is_fp8_block:
-            # Blackwell + FP8 block scales -> DeepGemm kernel
-            preferred = "deepgemm"
-        else:
-            # Default: Cutlass kernel
-            preferred = "moe_torch_flow"
-
-        # Check if preferred kernel is available in data, otherwise fallback
-        if self._wideep_moe_compute_data:
-            available_kernels = list(self._wideep_moe_compute_data.keys())
-            if preferred in available_kernels:
-                return preferred
-            elif available_kernels:
-                # Fallback to any available kernel
-                fallback = available_kernels[0]
-                logger.debug(f"Preferred MoE kernel '{preferred}' not available, falling back to '{fallback}'")
-                return fallback
-
-        return preferred
 
     def _get_value(self, data_value, metric: str = "latency"):
         """Thin wrapper — delegates to ``interpolation.get_value``."""
@@ -4056,220 +3939,25 @@ class PerfDatabase:
         database_mode: common.DatabaseMode | None = None,
         is_gated: bool = True,
     ) -> PerformanceResult | tuple[float, float, float]:
-        """
-        Query WideEP MoE compute latency (pure computation, excluding All2All communication).
+        """Delegates to ``TrtLLMWideEPMoE``; see
+        ``operations.moe.TrtLLMWideEPMoE._query_compute_table``."""
+        from aiconfigurator.sdk.operations.moe import TrtLLMWideEPMoE
 
-        This is for TensorRT-LLM WideEP scenarios with three EPLB modes:
-        - EPLB off: workload_distribution without "_eplb" suffix, num_slots = num_experts
-        - EPLB on: workload_distribution with "_eplb" suffix, num_slots = num_experts
-        - EPLB redundant: workload_distribution with "_eplb" suffix, num_slots > num_experts
-
-        The MoE computation kernel is automatically selected based on GPU architecture and quantization mode:
-        - SM >= 100 (Blackwell) with fp8_block -> DeepGemm kernel
-        - Otherwise -> Cutlass kernel (wideep_compute_cutlass)
-
-        Args:
-            num_tokens: Number of tokens
-            hidden_size: Hidden dimension size
-            inter_size: Intermediate size (FFN)
-            topk: Number of experts activated per token
-            num_experts: Total number of experts
-            num_slots: Number of expert slots (= num_experts for EPLB off/on, > num_experts for EPLB redundant)
-            moe_tp_size: MoE tensor parallelism size
-            moe_ep_size: MoE expert parallelism size
-            quant_mode: MoE quantization mode
-            workload_distribution: Workload distribution pattern (e.g., "power_law_1.01" or "power_law_1.01_eplb")
-            database_mode: Database mode (SILICON, EMPIRICAL, SOL, HYBRID)
-            is_gated: Whether MoE uses gated activation (SwiGLU=True, Relu2=False).
-                      Affects the number of GEMMs in SOL computation (3 for gated, 2 for non-gated).
-
-        Returns:
-            PerformanceResult: Latency in ms, energy accessible via .energy attribute.
-            For SOL_FULL mode: tuple of (sol_time, sol_math, sol_mem).
-        """
-
-        num_gemms = 3 if is_gated else 2
-
-        def get_sol(
-            num_tokens: int,
-            hidden_size: int,
-            inter_size: int,
-            topk: int,
-            num_experts: int,
-            num_slots: int,
-            moe_tp_size: int,
-            moe_ep_size: int,
-            quant_mode: common.MoEQuantMode,
-            workload_distribution: str,
-        ) -> tuple[float, float, float]:
-            """
-            Get the SOL (Speed of Light) time using Roofline model.
-
-            Uses num_slots instead of num_experts for weight memory calculation,
-            since WideEP EPLB redundant mode may replicate experts across slots.
-            """
-            total_tokens = num_tokens * topk
-            ops = total_tokens * hidden_size * inter_size * num_gemms * 2 // moe_ep_size // moe_tp_size
-            mem_bytes = quant_mode.value.memory * (
-                total_tokens // moe_ep_size * hidden_size * 2  # input+output
-                + total_tokens // moe_ep_size * inter_size * num_gemms // moe_tp_size  # intermediate activations
-                + hidden_size
-                * inter_size
-                * num_gemms
-                // moe_tp_size
-                * min(num_slots // moe_ep_size, total_tokens // moe_ep_size)  # weights (use num_slots)
-            )
-            sol_math = ops / (self.system_spec["gpu"]["bfloat16_tc_flops"] * quant_mode.value.compute) * 1000
-            sol_mem = mem_bytes / self.system_spec["gpu"]["mem_bw"] * 1000
-            sol_time = max(sol_math, sol_mem)
-            return sol_time, sol_math, sol_mem
-
-        def get_empirical_from_sol(
-            num_tokens: int,
-            hidden_size: int,
-            inter_size: int,
-            topk: int,
-            num_experts: int,
-            num_slots: int,
-            moe_tp_size: int,
-            moe_ep_size: int,
-            quant_mode: common.MoEQuantMode,
-            workload_distribution: str,
-        ) -> float:
-            """Get the empirical estimation: SOL / scale_factor."""
-            latency = get_sol(
-                num_tokens,
-                hidden_size,
-                inter_size,
-                topk,
-                num_experts,
-                num_slots,
-                moe_tp_size,
-                moe_ep_size,
-                quant_mode,
-                workload_distribution,
-            )[0]
-            scale_factor = 0.4
-            return latency / scale_factor
-
-        if database_mode is None:
-            database_mode = self._default_database_mode
-
-        if database_mode == common.DatabaseMode.SOL:
-            sol_latency = get_sol(
-                num_tokens,
-                hidden_size,
-                inter_size,
-                topk,
-                num_experts,
-                num_slots,
-                moe_tp_size,
-                moe_ep_size,
-                quant_mode,
-                workload_distribution,
-            )[0]
-            return PerformanceResult(sol_latency, energy=0.0, source="sol")
-        elif database_mode == common.DatabaseMode.SOL_FULL:
-            return get_sol(
-                num_tokens,
-                hidden_size,
-                inter_size,
-                topk,
-                num_experts,
-                num_slots,
-                moe_tp_size,
-                moe_ep_size,
-                quant_mode,
-                workload_distribution,
-            )
-        elif database_mode == common.DatabaseMode.EMPIRICAL:
-            emp_latency = get_empirical_from_sol(
-                num_tokens,
-                hidden_size,
-                inter_size,
-                topk,
-                num_experts,
-                num_slots,
-                moe_tp_size,
-                moe_ep_size,
-                quant_mode,
-                workload_distribution,
-            )
-            return PerformanceResult(emp_latency, energy=0.0, source="empirical")
-
-        # Automatically select MoE kernel based on GPU architecture and quant mode
-        kernel_source = self._select_moe_kernel(quant_mode)
-        logger.debug(f"query_wideep_moe_compute: auto-selected kernel_source='{kernel_source}'")
-
-        # SILICON or HYBRID mode - use database
-        def get_silicon():
-            self._wideep_moe_compute_data.raise_if_not_loaded()
-            # Find the best matching distribution
-            kernel_data = self._wideep_moe_compute_data[kernel_source]
-            available_distributions = list(kernel_data[quant_mode].keys())
-            if workload_distribution in available_distributions:
-                used_distribution = workload_distribution
-            else:
-                # Fallback: try to find a similar distribution or use the first available
-                used_distribution = available_distributions[0] if available_distributions else None
-                if used_distribution is None:
-                    raise KeyError(f"No distribution available for kernel={kernel_source}, quant_mode={quant_mode}")
-                logger.debug(f"Distribution '{workload_distribution}' not found, using '{used_distribution}' instead")
-
-            moe_dict = kernel_data[quant_mode][used_distribution][topk][num_experts][hidden_size][inter_size][
-                num_slots
-            ][moe_tp_size][moe_ep_size]
-
-            num_left, num_right = self._nearest_1d_point_helper(
-                num_tokens,
-                list(moe_dict.keys()),
-                inner_only=False,
-            )
-            result = self._interp_1d(
-                [num_left, num_right],
-                [moe_dict[num_left], moe_dict[num_right]],
-                num_tokens,
-            )
-
-            if isinstance(result, dict):
-                lat = result["latency"]
-                energy = result.get("energy", 0.0)
-            else:
-                lat = result
-                energy = 0.0
-
-            return self._interp_pr(lat, energy=energy)
-
-        def get_empirical() -> float:
-            # Simple empirical fallback based on SOL
-            total_tokens = num_tokens * topk
-            ops = total_tokens * hidden_size * inter_size * 3 * 2 // moe_ep_size // moe_tp_size
-            sol_math = ops / (self.system_spec["gpu"]["bfloat16_tc_flops"] * quant_mode.value.compute) * 1000
-            return sol_math / 0.4  # Empirical scale factor
-
-        return self._query_silicon_or_hybrid(
-            get_silicon=get_silicon,
-            get_empirical=get_empirical,
+        return TrtLLMWideEPMoE._query_compute_table(
+            self,
+            num_tokens=num_tokens,
+            hidden_size=hidden_size,
+            inter_size=inter_size,
+            topk=topk,
+            num_experts=num_experts,
+            num_slots=num_slots,
+            moe_tp_size=moe_tp_size,
+            moe_ep_size=moe_ep_size,
+            quant_mode=quant_mode,
+            workload_distribution=workload_distribution,
             database_mode=database_mode,
-            error_msg=(
-                f"Failed to query wideep moe compute data (kernel={kernel_source}) for "
-                f"{num_tokens=}, {hidden_size=}, {inter_size=}, {topk=}, {num_experts=}, "
-                f"{num_slots=}, {moe_tp_size=}, {moe_ep_size=}, {quant_mode=}, {workload_distribution=}"
-            ),
+            is_gated=is_gated,
         )
-
-    @staticmethod
-    def _normalize_alltoall_moe_quant_mode_for_table(
-        quant_mode: common.MoEQuantMode,
-    ) -> common.MoEQuantMode:
-        """
-        Normalize MoE quant modes for TRT-LLM alltoall perf table lookup.
-
-        `fp8_block` is a behavioral mode that reuses the `fp8` alltoall tables.
-        """
-        if quant_mode == common.MoEQuantMode.fp8_block:
-            return common.MoEQuantMode.fp8
-        return quant_mode
 
     @functools.lru_cache(maxsize=32768)
     def query_trtllm_alltoall(
@@ -4285,217 +3973,22 @@ class PerfDatabase:
         database_mode: common.DatabaseMode | None = None,
         moe_backend: str | None = None,
     ) -> PerformanceResult | tuple[float, float, float]:
-        """
-        Query TRT-LLM All2All communication latency.
+        """Delegates to ``TrtLLMWideEPMoEDispatch``; see
+        ``operations.moe.TrtLLMWideEPMoEDispatch._query_alltoall_table``."""
+        from aiconfigurator.sdk.operations.moe import TrtLLMWideEPMoEDispatch
 
-        Covers both WideEP (NVLinkTwoSided) and CutlassFusedMoE (NVLinkOneSided) paths.
-        The All2All communication method is automatically selected based on GPU architecture
-        and MoE backend type via _select_alltoall_kernel.
-
-        Args:
-            op_name: Operation name, one of:
-                - "alltoall_prepare": Prepare phase
-                - "alltoall_dispatch": Token dispatch phase
-                - "alltoall_combine": Result combine phase
-                - "alltoall_combine_low_precision": Low precision combine
-            num_tokens: Number of tokens
-            hidden_size: Hidden dimension size
-            topk: Number of experts activated per token
-            num_experts: Total number of experts
-            moe_ep_size: MoE expert parallelism size
-            quant_mode: MoE quantization mode
-            moe_backend: MoE backend identifier for kernel selection.
-                "wideep" -> NVLinkTwoSided;
-                "CUTLASS"/"TRTLLM"/None -> NVLinkOneSided;
-                "DEEPGEMM"/"CUTE_DSL" -> NotEnabled.
-            node_num: Number of nodes. If None, computed as moe_ep_size // 4
-            database_mode: Database mode
-
-        Returns:
-            PerformanceResult: Latency in ms, energy accessible via .energy attribute.
-
-        Raises:
-            ValueError: If op_name is not valid
-            PerfDataNotAvailableError: If alltoall perf data is unavailable for this version
-        """
-
-        def get_sol(
-            num_tokens: int,
-            hidden_size: int,
-            topk: int,
-            num_experts: int,
-            moe_ep_size: int,
-            quant_mode: common.MoEQuantMode,
-            node_num: int,
-        ) -> tuple[float, float, float]:
-            """
-            Get the SOL time for All2All communication.
-
-            All2All transfers token data between GPUs:
-            - prepare: lightweight metadata exchange (topk * 4 bytes per token)
-            - dispatch: each token sent once per unique remote rank (deduplication).
-              remote_ranks = min(topk, num_experts, ep_size - 1), bytes use quant_mode precision.
-            - combine: standard returns results in bfloat16 (2 B/elem);
-              low-precision variant returns results in fp4 (0.5 B/elem).
-              remote_ranks = min(topk, num_experts, ep_size - 1).
-            """
-            is_inter_node = node_num > 1
-
-            if is_inter_node:
-                bw = self.system_spec["node"]["inter_node_bw"]
-            else:
-                bw = self.system_spec["node"]["intra_node_bw"]
-
-            remote_ranks = min(topk, num_experts, moe_ep_size - 1)
-
-            if op_name == "alltoall_prepare":
-                data_bytes = num_tokens * topk * 4  # token routing indices, ~4 bytes per entry
-            elif "combine" in op_name:
-                bytes_per_element = 0.5 if "low_precision" in op_name else 2
-                data_bytes = num_tokens * remote_ranks * hidden_size * bytes_per_element
-            else:
-                # dispatch: per-rank deduplication, use quant_mode precision
-                data_bytes = num_tokens * remote_ranks * hidden_size * quant_mode.value.memory
-
-            sol_comm = data_bytes / bw * 1000  # ms
-            sol_time = sol_comm
-            return sol_time, sol_comm, 0.0
-
-        def get_empirical_from_sol(
-            num_tokens: int,
-            hidden_size: int,
-            topk: int,
-            num_experts: int,
-            moe_ep_size: int,
-            quant_mode: common.MoEQuantMode,
-            node_num: int,
-        ) -> float:
-            """Get the empirical estimation: SOL / scale_factor."""
-            latency = get_sol(
-                num_tokens,
-                hidden_size,
-                topk,
-                num_experts,
-                moe_ep_size,
-                quant_mode,
-                node_num,
-            )[0]
-            scale_factor = 0.5
-            return latency / scale_factor
-
-        if database_mode is None:
-            database_mode = self._default_database_mode
-
-        table_quant_mode = self._normalize_alltoall_moe_quant_mode_for_table(quant_mode)
-
-        # Compute node_num if not provided
-        if node_num is None:
-            if moe_ep_size < 4:
-                node_num = 1
-            else:
-                node_num = moe_ep_size // 4
-            logger.debug(f"query_trtllm_alltoall: node_num not specified, using {node_num} (moe_ep_size={moe_ep_size})")
-
-        valid_op_names = ["alltoall_prepare", "alltoall_dispatch", "alltoall_combine", "alltoall_combine_low_precision"]
-        if op_name not in valid_op_names:
-            raise ValueError(f"Invalid op_name '{op_name}'. Must be one of {valid_op_names}")
-
-        if database_mode == common.DatabaseMode.SOL:
-            sol_latency = get_sol(
-                num_tokens,
-                hidden_size,
-                topk,
-                num_experts,
-                moe_ep_size,
-                quant_mode,
-                node_num,
-            )[0]
-            return PerformanceResult(sol_latency, energy=0.0, source="sol")
-        elif database_mode == common.DatabaseMode.SOL_FULL:
-            return get_sol(
-                num_tokens,
-                hidden_size,
-                topk,
-                num_experts,
-                moe_ep_size,
-                quant_mode,
-                node_num,
-            )
-        elif database_mode == common.DatabaseMode.EMPIRICAL:
-            emp_latency = get_empirical_from_sol(
-                num_tokens,
-                hidden_size,
-                topk,
-                num_experts,
-                moe_ep_size,
-                quant_mode,
-                node_num,
-            )
-            return PerformanceResult(emp_latency, energy=0.0, source="empirical")
-
-        kernel_source = self._select_alltoall_kernel(quant_mode, moe_ep_size, topk, moe_backend=moe_backend)
-        logger.debug(
-            f"query_trtllm_alltoall: auto-selected kernel_source='{kernel_source}' (moe_backend={moe_backend})"
-        )
-
-        if kernel_source == "NotEnabled":
-            if database_mode == common.DatabaseMode.SOL_FULL:
-                return (0.0, 0.0, 0.0)
-            return PerformanceResult(0.0, energy=0.0, source="empirical")
-
-        # SILICON or HYBRID mode - use database
-        def get_silicon():
-            if not getattr(self, "_trtllm_alltoall_data", None):
-                raise PerfDataNotAvailableError(
-                    f"TRT-LLM alltoall perf data not available for version '{self.version}'. "
-                    "Use HYBRID or EMPIRICAL database mode."
-                )
-            self._trtllm_alltoall_data.raise_if_not_loaded()
-            kernel_data = self._trtllm_alltoall_data[kernel_source]
-            alltoall_dict = kernel_data[op_name][table_quant_mode][node_num][hidden_size][topk][num_experts][
-                moe_ep_size
-            ]
-
-            num_left, num_right = self._nearest_1d_point_helper(
-                num_tokens,
-                list(alltoall_dict.keys()),
-                inner_only=False,
-            )
-            result = self._interp_1d(
-                [num_left, num_right],
-                [alltoall_dict[num_left], alltoall_dict[num_right]],
-                num_tokens,
-            )
-
-            if isinstance(result, dict):
-                lat = result["latency"]
-                energy = result.get("energy", 0.0)
-            else:
-                lat = result
-                energy = 0.0
-
-            return self._interp_pr(lat, energy=energy)
-
-        def get_empirical() -> float:
-            return get_empirical_from_sol(
-                num_tokens,
-                hidden_size,
-                topk,
-                num_experts,
-                moe_ep_size,
-                quant_mode,
-                node_num,
-            )
-
-        return self._query_silicon_or_hybrid(
-            get_silicon=get_silicon,
-            get_empirical=get_empirical,
+        return TrtLLMWideEPMoEDispatch._query_alltoall_table(
+            self,
+            op_name=op_name,
+            num_tokens=num_tokens,
+            hidden_size=hidden_size,
+            topk=topk,
+            num_experts=num_experts,
+            moe_ep_size=moe_ep_size,
+            quant_mode=quant_mode,
+            node_num=node_num,
             database_mode=database_mode,
-            error_msg=(
-                f"Failed to query trtllm alltoall data for {op_name} (kernel={kernel_source}), "
-                f"moe_backend={moe_backend}, node_num={node_num}, {num_tokens=}, {hidden_size=}, "
-                f"{topk=}, {num_experts=}, {moe_ep_size=}, {quant_mode=}"
-            ),
+            moe_backend=moe_backend,
         )
 
     # ═══════════════════════════════════════════════════════════════════
